Update translations for multiple languages to enhance API endpoint and filter management
This commit updates the language files for Czech, German, English, Spanish, French, Italian, and Japanese, adding new keys and translations related to API endpoints, filters, and error messages. The changes improve the user experience by providing clearer instructions and feedback in various languages, ensuring consistency across the application.
